data_foundation_project/requirements.txt
geopandas
fiona
shapely
pyproj
pandas
fastapi
uvicorn
jupyterlab
snowflake-connector-python>=3.6.0
snowflake-sqlalchemy>=1.5.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.5.0
httpx>=0.25.0
